delete recipe by owner only

+++ recipes/post_delete.php
@@ -2,6 +2,7 @@
 
 // load all data to be used 
 include_once($_SERVER['DOCUMENT_ROOT'] . "/config/mysql.php");
+include_once($_SERVER['DOCUMENT_ROOT'] .  '/config/user.php');
 include_once($_SERVER['DOCUMENT_ROOT'] . "/variables/variables.php");
 
 $recipe_id = $_POST['id'];
@@ -12,10 +13,23 @@ if (!isset($recipe_id) || (!is_numeric($recipe_id)))
     return;
 }	
 
-$deleteRecipe = $mysqlClient->prepare('DELETE FROM recipes WHERE recipe_id = :recipe_id');
+// delete the recipe where the id matches & user is owner, who is logged in.
+$deleteRecipe = $mysqlClient->prepare('DELETE FROM recipes WHERE recipe_id = :recipe_id AND author = :author');
 $deleteRecipe->execute([
   'recipe_id' => $recipe_id,
+  'author' => $loggedUser['email'],
 ]);
 
 header('Location: '.$rootUrl.'index.php');
 ?>
+
+
+<!-- // EXAMPLE 3
+  // $sqlQuery ="SELECT * FROM `recipes` WHERE author = :author AND is_enabled = :is_enabled";
+  // $recipesStatement = $db->prepare($sqlQuery);
+  // $recipesStatement->execute([
+  //   'author' => $_SESSION['LOGGED_USER'],
+  //   'is_enabled' => true,
+  // ]);
+
+  // $recipes = $recipesStatement->fetchAll(); -->
